bugfix: 插件重试会导致构建重试出错 #556
