+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/service/PipelineService.kt
@@ -76,6 +76,7 @@ import com.tencent.devops.process.service.view.PipelineViewService
 import com.tencent.devops.process.utils.PIPELINE_VIEW_ALL_PIPELINES
 import com.tencent.devops.process.utils.PIPELINE_VIEW_FAVORITE_PIPELINES
 import com.tencent.devops.process.utils.PIPELINE_VIEW_MY_PIPELINES
+import com.tencent.devops.process.utils.PipelineVarUtil
 import com.tencent.devops.store.api.common.ServiceStoreResource
 import org.jooq.DSLContext
 import org.jooq.Record
@@ -607,10 +608,34 @@ class PipelineService @Autowired constructor(
                 errorCode = ProcessMessageCode.ERROR_PIPELINE_MODEL_NOT_EXISTS,
                 defaultMessage = "指定要复制的流水线-模型不存在")
         try {
-            val buildNo = (model.stages[0].containers[0] as TriggerContainer).buildNo
+            val triggerContainer = model.stages[0].containers[0] as TriggerContainer
+            val buildNo = triggerContainer.buildNo
+            val params = triggerContainer.params
             if (buildNo != null) {
                 buildNo.buildNo = pipelineRepositoryService.getBuildNo(projectId = projectId, pipelineId = pipelineId) ?: buildNo.buildNo
             }
+            val newParams = mutableListOf<BuildFormProperty>()
+            params.forEach {
+                // 变量名从旧转新: 兼容从旧入口写入的数据转到新的流水线运行
+                val newVarName = PipelineVarUtil.oldVarToNewVar(it.id)
+                if (!newVarName.isNullOrBlank()) {
+                    newParams.add(BuildFormProperty(
+                        id = newVarName!!,
+                        required = it.required,
+                        type = it.type,
+                        defaultValue = it.defaultValue,
+                        options = it.options,
+                        desc = it.desc,
+                        repoHashId = it.repoHashId,
+                        relativePath = it.relativePath,
+                        scmType = it.scmType,
+                        containerType = it.containerType,
+                        glob = it.glob,
+                        properties = it.properties
+                    ))
+                } else newParams.add(it)
+            }
+            triggerContainer.params = newParams
 
             // 获取流水线labels
             val groups = pipelineGroupService.getGroups(userId = userId, projectId = projectId, pipelineId = pipelineId)
